Sunday, 2 October 2016 Cumulative update Hi everyone,
it's been a while, my summer was packed and I didn't have much time or energy to work on TacO. Meanwhile interest in the overlay seems to have picked up, and many of you have started creating marker packs.
This release is more of a maintenance build addressing some issues that popped up with the recent Windows 10 anniversary update and some small quality of life feature requests that came in. It's also a step for me to get back to spending a bit more time on the overlay in the coming months. I have some pretty interesting ideas :)

The Xera encounter will now have an HP bar grid, and numbered markers for the rifts to make coordinating easier.

There was also an anonymous donation of markers for Ember Bay containing Mursaat Token and Tome locations. I have validated and fixed this pack and included it in this release.

You can download the new build here.

Change notes:

    The TS3 overlay will now put anyone who talks to the top of the list. This will make it a lot more useful for large channels.
    Tiny display fix for the TS3 overlay list (last item won't be partially displayed anymore)
    Overhauled the way TacO is placed over the GW2 window to be less intrusive.
    TacO will no longer show up as a separate program on the taskbar or the alt-tab menu.
    Integrated AngelScript for future use.
    Fixed a TS3 overlay issue that would sometimes happen if TS3 was started after TacO.
    TacO will no longer show up on the GW2 login window.
    Added HP bar grid for Xera.
    Added numbered markers for Xera rifts.
    Added Ember Bay collectible markers for Mursaat Tokens and Tomes, courtesy of an Anonymous comment on the blog.

Step: 14

[A ] new file:

	Data/arrow.png
	Data/circle.png
	Data/cross.png
	Data/heart.png
	Data/spiral.png
	Data/square.png
	Data/star.png
	Data/textbox.png
	Data/triangle.png
	POIs/Data/Rising_Flames.png
	POIs/collectibles_mursaattokens.xml
	POIs/collectibles_tabletsoftheunseen.xml

[ M] modified - Changes not staged for commit:

	GW2TacO.exe
	GW2TacO.pdb
	POIs/raid1_stronghold_markers.xml
	UI.wbs
	UI.xml
	UI_large.css
	UI_larger.css
	UI_normal.css
	UI_small.css
	hpgrids.xml
	locationaltimers.xml
	maptimer.xml