Wednesday, 19 October 2016 Trick or Treat! Hi!
Quick update here to join in on the Halloween fun.
I marked all the carving pumpkins in Lions Arch and the Mad King's Labyrinth.
A new feature has also been added in according to this, there's a new marker behavior type that will remember the instance the marker was triggered on. This means that if you trigger markers of this type and join another instance of the map, all the markers will be up again for you to run through them. If you manage to get back into the same instance you were in before, you'll see where you haven't been yet.
This feature is experimental and rough though, so don't start using it yet for Bloodstone Crystal nodes etc :)

You can download the new build here.

Change notes

    Added markers for carving pumkins in Lion's Arch and the Mad king's Labyrinth.
    Added marker behavior type 6: OncePerInstance.
    Removed some logging that might have flooded the log file before GW2 was running.

Step: 1

[A ] new file:

	POIs/Data/Halloween.png
	POIs/Festival_Halloween_CarvingPumpkins.xml
	notepad.txt

[ M] modified - Changes not staged for commit:

	GW2TacO.exe
	GW2TacO.pdb