Friday, 11 November 2016 Hoppity-hop Hi!
I just got a big surprise package in the mail from a group of guys who took a liking to TacO - they went around ALL the jumping puzzles in the game and marked the routes through them. The result is a marker pack of about 5000 markers.
There's no clutter though like with SAB: only a single marker is shown at a time, marking the place you should go next. The starting points of the puzzles are all marked separately as well.

So give it up to this awesome and dedicated team:

    All PvE-JP-Guides made by "Cloud Strife.7492"
    The Mists and Fractal JP-Guides made by "Destiny FNC.5018"
    Rejumped, adjusted and sorted by "Xenophy.2716"

The new build contains this huge marker pack now. If anyone finds problems with it, they can contact the guys at xenotaco@web.de, and they'll get me an updated version to release.

In this release I also recategorized and reintroduced all the previous festival markers. This means the carving pumpkins, wintersday orphans, and new year's fireworks will become a permanent package in the overlay.

No real feature updates for now since I'm busy with work and other stuff. I'm slowly working on the core of the "next big feature" to add to the overlay, however to truly make it shine I'll need someone good at writing and telling a story and familiar with Tyria and/or GW2 lore. Anyone interested out there? Maybe from the RP community?

You can download the new build from here.

Change log:

    If the GW2 UI doesn't tick for a third of a second, TacO will be hidden. This will make loading screens and video cutscenes TacO-free.
    Added a huge Jumping Puzzle marker pack courtesy of a very dedicated team of players
    Added back the orphan markers in anticipation of the upcoming wintersday
    Rearranged some of the marker categories (festival markers are now grouped together)

Step: 8

[A ] new file:

	POIs/Data/Battle_bw.png
	POIs/Data/jpguide.png
	POIs/Data/jpstart.png
	POIs/Data/one.png
	POIs/Data/portlabconsole1.png
	POIs/Data/portlabconsole2.png
	POIs/Data/three.png
	POIs/Data/two.png
	POIs/Festival_CanthanNewYear_Firecrackers.xml
	POIs/Festival_Wintersday_Orphans.xml
	POIs/JP_Megapack.xml

[ M] modified - Changes not staged for commit:

	GW2TacO.exe
	GW2TacO.pdb
	POIs/Adventure_FloorIsLava.xml
	locationaltimers.xml